third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/unloading-documents/005-expected.txt

This is a testharness.js-based test.
[FAIL] document.open in pagehide in iframe
  assert_equals: expected "012345678" but got "0123567"
Harness: the test ran to completion.
